In p:file-copy, don’t delete the originals!
Fix #28

+++ xmlcalabash/src/main/kotlin/com/xmlcalabash/steps/file/FileCopyOrMove.kt
@@ -89,8 +89,10 @@ abstract class FileCopyOrMove(stepType: QName): FileStep(stepType) {
                 }
             }
 
-            if (source.isDirectory) {
-                source.deleteRecursively()
+            if (stepType == NsP.fileMove) {
+                if (source.isDirectory) {
+                    source.deleteRecursively()
+                }
             }
         }
